// src/app/contact/contact.interface.ts
export interface Contact {
    TableRight: FieldRightOrTableRight;
    FieldProperties: FieldProperties;
    ContactId: number;
    Name: string;
    Department: string;
    OrgNr: string;
    Number1: string;
    Number2: string;
    UpdatedDate: string;
    CreatedDate: string;
    Emails?: (EmailsEntityOrUrlsEntityOrPhonesEntity)[] | null;
    Interests?: (InterestsEntity)[] | null;
    Urls?: (EmailsEntityOrUrlsEntityOrPhonesEntity)[] | null;
    Phones?: (EmailsEntityOrUrlsEntityOrPhonesEntity)[] | null;
    Faxes?: (null)[] | null;
    Description: string;
    UpdatedBy?: null;
    CreatedBy: CreatedBy;
    Associate?: null;
    Business: BusinessOrCategoryOrTicketPriorityOrCustomerLanguage;
    Category: BusinessOrCategoryOrTicketPriorityOrCustomerLanguage;
    Country: Country;
    Persons?: (null)[] | null;
    NoMailing: boolean;
    Kananame: string;
    Xstop: boolean;
    ActiveInterests: number;
    GroupId: number;
    ActiveStatusMonitorId: number;
    SupportAssociate?: null;
    TicketPriority: BusinessOrCategoryOrTicketPriorityOrCustomerLanguage;
    CustomerLanguage: BusinessOrCategoryOrTicketPriorityOrCustomerLanguage;
    Deleted: number;
    DbiAgentId: number;
    DbiLastSyncronized: string;
    DbiKey: string;
    DbiLastModified: string;
    SupportPerson?: null;
    Address: Address;
    Source: number;
    ActiveErpLinks: number;
    UserDefinedFields: FieldPropertiesOrUserDefinedFieldsOrExtraFields;
    ExtraFields: FieldPropertiesOrUserDefinedFieldsOrExtraFields;
  }
  export interface FieldRightOrTableRight {
    Mask: string;
    Reason: string;
  }
// ...
  export interface EmailsEntityOrUrlsEntityOrPhonesEntity {
    TableRight?: null;
    FieldProperties: FieldPropertiesOrUserDefinedFieldsOrExtraFields;
    Value: string;
    StrippedValue: string;
    Description: string;
  }
  export interface FieldPropertiesOrUserDefinedFieldsOrExtraFields {
  }
  export interface InterestsEntity {
    TableRight?: null;
    FieldProperties: FieldPropertiesOrUserDefinedFieldsOrExtraFields;
    Id: number;
    Name: string;
    ToolTip: string;
    Deleted: boolean;
    Rank: number;
    Type: string;
    ColorBlock: number;
    IconHint: string;
    Selected: boolean;
    LastChanged: string;
    ChildItems?: (null)[] | null;
    ExtraInfo: string;
    StyleHint: string;
    Hidden: boolean;
    FullName?: null;
  }
  export interface CreatedBy {
    TableRight: FieldRightOrTableRight;
    FieldProperties: FieldPropertiesOrUserDefinedFieldsOrExtraFields;
    AssociateId: number;
    Name: string;
    PersonId: number;
    Rank: number;
    Tooltip: string;
    Type: string;
    GroupIdx: number;
    FullName: string;
    FormalName: string;
    Deleted: boolean;
    EjUserId: number;
  }
  export interface BusinessOrCategoryOrTicketPriorityOrCustomerLanguage {
    TableRight?: null;
    FieldProperties: FieldPropertiesOrUserDefinedFieldsOrExtraFields;
    Id: number;
    Value: string;
    Tooltip: string;
  }
  export interface Country {
    TableRight?: null;
    FieldProperties: FieldPropertiesOrUserDefinedFieldsOrExtraFields;
    CountryId: number;
    Name: string;
    CurrencyId: number;
    EnglishName: string;
    TwoLetterISOCountry: string;
    ThreeLetterISOCountry: string;
    ImageDescription?: null;
    OrgNrText: string;
    InterAreaPrefix: string;
    DialInPrefix: string;
    ZipPrefix: string;
    DomainName: string;
    AddressLayoutId: number;
    DomesticAddressLayoutId: number;
    ForeignAddressLayoutId: number;
    Rank: number;
    Tooltip: string;
    Deleted: boolean;
  }
  export interface Address {
    TableRight?: null;
    FieldProperties: FieldPropertiesOrUserDefinedFieldsOrExtraFields;
    Wgs84Latitude: number;
    Wgs84Longitude: number;
    LocalizedAddress?: ((EntityOrLocalizedAddressEntityEntity)[] | null)[] | null;
  }
  export interface EntityOrLocalizedAddressEntityEntity {
    TableRight?: null;
    FieldProperties: FieldPropertiesOrUserDefinedFieldsOrExtraFields;
    Name: string;
    Value: string;
    Tooltip: string;
    Label: string;
    ValueLength: number;
    AddressType: string;
  }